requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.0.0
psycopg2-binary==2.9.3  # PostgreSQL adapter
Flask-Bcrypt==0.7.1  # For hashing passwords
PyJWT==2.3.0  # JSON Web Tokens
